service for admin is ready.

+++ Product-Service/src/main/java/com/TheCoderKushagra/service/CategoryService.java
@@ -8,17 +8,22 @@ import com.TheCoderKushagra.entity.Category;
 import com.TheCoderKushagra.entity.CategoryAttribute;
 import com.TheCoderKushagra.repository.CategoryAttributeRepository;
 import com.TheCoderKushagra.repository.CategoryRepository;
+import com.TheCoderKushagra.repository.ProductAttributeRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Slf4j
 @Service
 public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
     @Autowired
     private CategoryAttributeRepository categoryAttributeRepository;
+    @Autowired
+    private ProductAttributeRepository productAttributeRepository;
 
     public CategoryResponse createCategory(CategoryRequest request) {
         Category myCategoryName = categoryRepository.findByCategoryName(request.getParentCategory());
@@ -35,19 +40,25 @@ public class CategoryService {
                 .build();
     }
 
-    public CategoryResponse seeAllCategory() {
-        return null;
+    public List<Category> seeAllCategory() {
+        return categoryRepository.findAll();
     }
 
-    public CategoryResponse seeCategoryById() {
-        return null;
-    }
+    public CategoryResponse upgradeCategory(String Id, CategoryRequest request) {
+        Category myCategoryName = categoryRepository.findByCategoryName(request.getParentCategory());
+        Category category = categoryRepository.findById(Id)
+                .orElseThrow(() -> new RuntimeException(Id + "is invalid!!"));
 
-    public CategoryResponse upgradeCategory() {
-        return null;
-    }
+        category.setCategoryName(request.getCategoryName());
+        category.setParentCategory(myCategoryName);
 
-    public void deleteCategory() {
+        Category saved = categoryRepository.save(category);
+        return CategoryResponse.builder()
+                .categoryId(saved.getCategoryId())
+                .categoryName(saved.getCategoryName())
+                .parentCategory(saved.getParentCategory())
+                .attributes(saved.getAttributes())
+                .build();
     }
 
     public CategoryAttributeResponse createCategoryAttribute(CategoryAttributeRequest request) {
@@ -67,22 +78,20 @@ public class CategoryService {
                 .build();
     }
 
-    public CategoryAttributeResponse updateCategoryAttribute() {
-        return null;
-    }
-
     public void DeleteCategoryAttribute(String id) {
         boolean exists = categoryAttributeRepository.existsById(id);
         if(exists) {
             CategoryAttribute categoryAttribute = categoryAttributeRepository.findById(id)
                     .orElseThrow(() -> new RuntimeException("Entity Don't exist"));
-            Category category = categoryAttribute.getCategory();
-
-            category.getAttributes().remove(categoryAttribute); // removed from category List
-            categoryAttributeRepository.deleteById(id); // removed from CategoryAttribute Table
-        } else {
-            throw new RuntimeException(id + "is invalid!!");
-        }
+            boolean inUse = productAttributeRepository.existsByAttribute(categoryAttribute);
+            if (inUse) {
+                log.error("Cannot delete category; Some products is using this attributes");
+            } else {
+                Category category = categoryAttribute.getCategory();
+                category.getAttributes().remove(categoryAttribute); // removed from category List
+                categoryAttributeRepository.deleteById(id); // removed from CategoryAttribute Table
+            }
+        } else { log.error("{} id invalid", id); }
     }
 
 }
